refactor: separate UserItem and NoResult component
